refactor(memory): split hardening_tests.rs (#1266)
hardening_tests.rs 726 -> 365 LOC; new hardening_tests_extra.rs (~397).
